refactor: split onboard web search probe
